streamlit-app/requirements.txt
streamlit==1.29.0
streamlit-webrtc==0.47.1
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.1.0
requests==2.31.0
pandas==2.1.3
plotly==5.18.0
python-dotenv==1.0.0
